mobile/android/tests/browser/chrome/test_migrate_ui.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1154504
Migrated from Robocop: https://bugzilla.mozilla.org/show_bug.cgi?id=1184186
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1154504</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="head_search.js"></script>
  <script type="application/javascript">

    Services.prefs.clearUserPref("browser.migration.version");
    Services.prefs.setBoolPref("nglayout.debug.paint_flashing", true);

    addTestEngines([
      { name: "bacon", details: ["", "bacon", "Search Bacon", "GET",
                                 "http://www.bacon.moz/?search={searchTerms}"] },
    ]).then(engines => {
      Services.prefs.setCharPref("browser.search.defaultenginename", engines[0].name);

      let BrowserApp = Services.wm.getMostRecentWindow("navigator:browser").BrowserApp;

      // This performs the serach migration asynchronously, but the search service is already initialized
      // by `addTestEngines`, so we don't need to worry about waiting before checking the new engine.
      BrowserApp._migrateUI();

      // Check that migration version increased.
      is(Services.prefs.getIntPref("browser.migration.version"), 3, "found expected version");

      // Check that user pref value was reset.
      is(Services.prefs.prefHasUserValue("nglayout.debug.paint_flashing"), false, "found expected user value");

      function searchObserver(s, t, d) {
        Services.obs.removeObserver(searchObserver, "default-search-engine-migrated");
        is(Services.search.defaultEngine.name, engines[0].name, "found expected default search engine");
      }
      Services.obs.addObserver(searchObserver, "default-search-engine-migrated", false);
    });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1154504">Mozilla Bug 1154504</a>
<br>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1184186">Migrated from Robocop testMigrateUI</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
